+++ cmd/libs/go2idl/client-gen/test_apis/testgroup.k8s.io/v1/register.go
